+++ R/autocorr.R
@@ -1,31 +1,185 @@
-
+#' Compute the standard error accounting for empirical autocorrelations
+#'
+#' @param x numeric vector
+#' @param na.rm logical. Should missing values be removed?
+#' @param effCor numeric vector of effective correlation components
+#'  first entry at zero lag equals one. See \code{\link{computeEffectiveAutoCorr}}
+#' @param effCov alternative to specifying effCor: numeric vector of 
+#'  effective covariance components
+#'  first entry is the variance. See \code{\link{computeEffectiveAutoCorr}}
+#' @param nEff possibility to specify precomputed number of effective 
+#'  observations for speedup.
+#'  
+#' @details The default uses empirical autocorrelation
+#'  estimates from the supplied data up to first negative component.
+#'  For short series of \code{x} it is strongly recommended to to
+#'  provide \code{effCov} that was estimated on a longer time series.
+#'
 #' @export
-computeEffectiveNumObs <- function(
-  ### Compute the effective number of observations taking into account autocorrelation
-  res  ##<< numeric of autocorrelated numbers, usually observation - model residuals
-  , effAcf = computeEffectiveAutoCorr(res) ##<< autocorrelation coefficients.
-  ## The first entry is fixed at 1 for zero distance.
-  ## May provide precomputed for efficiency or computed from a larger sample.
-  , na.rm = FALSE  ##<< a logical value indicating whether NA values should be 
-  ## stripped before the computation proceeds. 
+#' @return numeric scalar of standard error of the mean of x
+seCor <- function(
+  x  
+  , effCor = if (missing(effCov)) computeEffectiveAutoCorr(x) else 
+      effCov/var(x, na.rm = TRUE)
+  , na.rm = FALSE 
+  , effCov # compatibility to seCor up to version 0.1.8
+  , nEff = computeEffectiveNumObs(x, effCor, na.rm = na.rm)
 ){
-  ##references<< 
-  ## \code{Zieba & Ramza (2011) 
-  ## Standard Deviation of the Mean of Autocorrelated 
-  ## Observations Estimated with the Use of the Autocorrelation Function 
-  ## Estimated From the Data. 
-  ## Metrology and Measurement Systems, 
-  ## Walter de Gruyter GmbH, 18 10.2478/v10178-011-0052-x}
-  ## 
-  ## \code{Bayley & Hammersley (1946) 
-  ## The "effective" number of independent observations in an autocorrelated 
-  ## time series. 
-  ## Supplement to the Journal of the Royal Statistical Society, JSTOR,8,184-197}
-  #
-  ##details<< Handling of NA values: NAs at the beginning or end and are 
-  ## just trimmed before computation and pose no problem. 
-  ## However with NAs aside from edges, the return value is biased low,
-  ## because correlation terms are subtracted for those positions.
+  n <- if (na.rm) length(na.omit(x)) else length(x)
+  if (n < 2) return(NA_real_)
+  if (var(x, na.rm = TRUE) == 0) return(0)
+  varCorVal <- varCor(x, nEff = nEff, na.rm = TRUE)/nEff
+  sqrt(varCorVal)
+}
+
+#' Compute the unbiased variance accounting for empirical autocorrelations
+#'
+#' @param x numeric vector
+#' @param na.rm logical. Should missing values be removed?
+#' @param effCor numeric vector of effective correlation components
+#'  first entry at zero lag equals one. See \code{\link{computeEffectiveAutoCorr}}
+#'  The effective correlation is passed to \code{\link{computeEffectiveNumObs}}.
+#' @param nEff possibility to specify precomputed number of effective 
+#'  observations for speedup.
+#'  
+#' @details The default uses empirical autocorrelation
+#'  estimates from the supplied data up to first negative component.
+#'  For short series of \code{x} it is strongly recommended to to
+#'  provide \code{effCov} that was estimated on a longer time series.
+#'
+#' @export
+#' @return numeric scalar of unbiased variation of x
+varCor <- function(
+  x  
+  , effCor = computeEffectiveAutoCorr(x) 
+  , na.rm = FALSE 
+  , nEff = computeEffectiveNumObs(x, effAcf = effCor)
+){
+  var_uncorr = var(x, na.rm=na.rm) 
+  if (!is.finite(var_uncorr) || (var_uncorr == 0)) return(var_uncorr)
+  n = length(x)
+  if (n <= 1) return(var_uncorr)
+  # BLUE Var(x) for correlated: Zieba11 eq.(1) 
+  nmiss = sum(is.na(x))
+  nfin = n - nmiss
+  var_uncorr*(nfin-1)*nEff/(nfin*(nEff-1))
+}
+
+seCor_depr2103 <- function(
+  x  
+  , na.rm = FALSE 
+  , effCov = computeEffectiveAutoCorr(x, type = "covariance")
+){
+  n <- if (na.rm) length(na.omit(x)) else length(x)
+  #effAcf <- computeEffectiveAutoCorr(x, na.action = {if(na.rm) na.omit else na.pass})
+  # do not remove NAs for autocorrelation computation to preserve distances
+  if (n == 0) return(NA_real_)
+  g1 <- effCov[1:min(length(x),length(effCov))]
+  kmax <- length(g1) - 1
+  # if there is no empirical autocorrelation
+  if (kmax == 0) {
+    varx <- var(x, na.rm = na.rm)
+    return(sqrt(varx/n))
+  } 
+  g0 <- g1[1]
+  g <- g1[-1]
+  k <- 1:kmax
+  # note that we use autocovariance instead of autocorrelation
+  varCorB <- (g0 + 2*sum( (n - k)/n * g))/n
+  sqrt(varCorB)
+}
+
+
+#' Compute the effective number of observations taking into account autocorrelation
+#' 
+#' @param res numeric of autocorrelated numbers, usually observation -
+#'  model residuals
+#' @param effAcf autocorrelation coefficients.
+#'  The first entry is fixed at 1 for zero distance.
+#' @param na.rm if not set to TRUE will return NA in there are missings
+#'  in the series
+#' @param exact.na if set to FALSE then do not count and correct for missing
+#'   in the sum of autocorrelation terms. This is faster, but results are
+#'   increasingly biased high with increasing number of missings. 
+#'
+#' @references 
+#' \code{Zieba & Ramza (2011) 
+#' Standard Deviation of the Mean of Autocorrelated 
+#' Observations Estimated with the Use of the Autocorrelation Function 
+#' Estimated From the Data. 
+#' Metrology and Measurement Systems, 
+#' Walter de Gruyter GmbH, 18 10.2478/v10178-011-0052-x}
+#'
+#' \code{Bayley & Hammersley (1946) 
+#' The "effective" number of independent observations in an autocorrelated 
+#' time series. 
+#' Supplement to the Journal of the Royal Statistical Society, JSTOR,8,184-197}
+#'
+#' @details Assumes records of all times present. 
+#' DO NOT REMOVE OR FILTER NA records before. 
+#' The length of the time series is used.
+#' @details Handling of NA values: The formula from Zieba 2011 is extended
+#' to subtract the number of missing pairs in the count of correlation terms.
+#' If `exact.na=false` the original formula is used (after trimming edge-NAs).
+#' 
+#' @export
+#' @return integer scalar: effective number of observations
+#' @examples
+#' # generate autocorrelated time series
+#' res <- stats::filter(rnorm(1000), filter = rep(1,5), circular = TRUE)
+#' res[100:120] <- NA
+#' # plot the series of autocorrelated random variables
+#' plot(res)
+#' # plot their empirical autocorrelation function
+#' acf(res, na.action = na.pass)
+#' #effAcf <- computeEffectiveAutoCorr(res)
+#' # the effective number of parameters is less than number of 1000 samples
+#' (nEff <- computeEffectiveNumObs(res, na.rm = TRUE))
+computeEffectiveNumObs <- function(
+  res  
+  , effAcf = computeEffectiveAutoCorr(res)
+  , na.rm = FALSE 
+  , exact.na = TRUE
+){
+  if (!isTRUE(na.rm) & any(is.na(res))) return(NA_integer_)
+  resTr <- .trimNA(res)
+  lacf = length(effAcf) -1 # acf starts with lag 0
+  isFin <- is.finite(resTr)
+  if (lacf < 1) return(sum(isFin))
+  n <- length(resTr)
+  if (n < 2) return(n)
+  k = 1:min(n-1,lacf) 
+  if (exact.na) {
+    # number of missing combinations due to missing in x
+    mka = count_NA_forlags(resTr, 0:length(k))
+    m0 = mka[1]
+    mk = mka[-1]
+    nf = n - m0
+    neff = nf/(1 + 2/nf*sum((n - k -mk) * effAcf[k+1]))  
+  } else {
+    neff = n/(1 + 2/n*sum((n - k) * effAcf[k+1]))  
+  }
+  neff
+}
+
+count_NA_forlags <- function(x, lags=0:(length(x)-1)) {
+  lx = length(x)
+  xb = is.na(x)
+  cntna0 = sum(xb)
+  vapply(lags, function(lag) {
+    if (lag == 0) return(cntna0)
+    if (lag > (lx-1)) stop(
+      "expected lag < length(x)-1=",lx-1," but got lag ",lag)
+    xm = cbind(xb[1:(lx-lag)],xb[-(1:lag)])
+    sum(rowSums(xm) > 0)
+  },0)
+}
+
+computeEffectiveNumObs_depr2103 <- function(
+  res  
+  , effAcf = computeEffectiveAutoCorr(res)
+  , na.rm = FALSE  
+){
   resTr <- .trimNA(res)
   if (!isTRUE(na.rm) & any(is.na(resTr))) return(NA_integer_)
   isFin <- is.finite(resTr)
@@ -34,99 +188,74 @@ computeEffectiveNumObs <- function(
   effAcfD <- effAcf[-1] # without zero distance
   # correlations may have been computed on a sample larger than given
   # then only use the 1..(n-1) components
+  # If there an NAs in resTr then nFin < length(resTr) and the numerator is 
+  # decreased.
+  # But autocorrelation is still accounted for pairs involving NA.
+  # Therefore, the denominator is overestimated and hence the fraction nEff is 
+  # underestimated.
   nC <- min(length(resTr) - 1, length(effAcfD))
+  #?nC <- min(n - 1, length(effAcfD))
   if (nC == 0) return(n)
   nEff0 <- n/(1 + 2*sum((1 - 1:nC/length(resTr))*effAcfD[1:nC]))
-  ##details<< Because of NA correlation terms, the computed effective number of
-  ## observations can be smaller than 1. In this case 1 is returned.
+  # alternative for missings: using n=nFin instead of length(resTr) in k/n
+  # i.e. computed on vector omitting NA
+  # k/nFin is smaller then k/length(resTr) -> (1-k/N) larger -> 
+  #   denominator larger -> nEff smaller 
+  # this increases the bias underestimating nEff -> prefer length(nTr) solution
   nEff <- max(1, nEff0)  
   if ( nEff > n) stop("encountered nEff larger than finite records.")
-  ##value<< integer scalar: effective number of observations
   nEff
 }
-attr(computeEffectiveNumObs,"ex") <- function(){
-  # generate autocorrelated time series
-  res <- stats::filter(rnorm(1000), filter = rep(1,5), circular = TRUE)
-  res[100:120] <- NA
-  # plot the series of autocorrelated random variables
-  plot(res)
-  # plot their empirical autocorrelation function
-  acf(res, na.action = na.pass)
-  #effAcf <- computeEffectiveAutoCorr(res)
-  # the effective number of parameters is less than number of 1000 samples
-  (nEff <- computeEffectiveNumObs(res, na.rm = TRUE))
-}
 
 
-.trimNA <- function(
-  ### remove NA values at the start and end
-  x  ##<< numeric vectpr
-){
-  nisna <- complete.cases(x)
-  idx <- cumsum(nisna > 0) & rev(cumsum(rev(nisna))) > 0
-  ##value<< subset of x with leading and trailing NAs removed
-  x[idx]
-}
-
+#' Estimate vector of effective components of the autocorrelation
+#' 
+#' @param res  numeric of autocorrelated numbers, usually observation - 
+#'  model residuals
+#' @param type type of residuals (see \code{\link{acf}})
+#'
+#' @details
+#'  Returns all components before first negative autocorrelation
+#' @references
+#'  \code{Zieba 2011 Standard Deviation of the Mean of Autocorrelated
+#'  Observations Estimated with the Use of the Autocorrelation Function Estimated
+#'  From the Data}
 #' @export
+#' @return numeric vector: strongest components of the autocorrelation function
+#' @examples
+#' # generate autocorrelated time series
+#' res <- stats::filter(rnorm(1000), filter = rep(1,5), circular = TRUE)
+#' res[100:120] <- NA
+#' (effAcf <- computeEffectiveAutoCorr(res))
 computeEffectiveAutoCorr <- function(
-  ### Return the vector of effective components of the autocorrelation
-  res  ##<< numeric of autocorrelated numbers, usually observation - model residuals
-  , type = "correlation"
-){
+  res, type = "correlation" ){
   # first compute empirical autocorrelations
   ans <- acf(res, na.action = na.pass, plot = FALSE, type = type)
   # next get the number of elements before crossing the zero line
   nC <- suppressWarnings(min(which(ans$acf <= 0)) - 1)
   if (!is.finite(nC)) {
-    # if there was no below zero correlation within defalt lag.max then
-    # repeat acf with compting all lags
+    # if there was no below zero correlation within default lag.max then
+    # repeat acf with computing all lags
     ans <- acf(res, na.action = na.pass, plot = FALSE, type = type, lag.max = Inf)
-    # append -1 so that nC equals to full lenght if no negative correlation
+    # append -1 so that nC equals to full length if no negative correlation
     nC <- min(which(c(ans$acf,-1) <= 0)) - 1
   }
-  ##details<<
-  ## Returns all components before first negative autocorrelation
-  ##references<<
-  ## \code{Zieba 2011 Standard Deviation of the Mean of Autocorrelated
-  ## Observations Estimated with the Use of the Autocorrelation Function Estimated
-  ## From the Data}
-  ##value<< numeric vector: strongest components of the autocorrelation function
   nC <- pmax(1,nC) # return at least one component
   ans$acf[1:nC]
 }
-attr(computeEffectiveAutoCorr,"ex") <- function(){
-  # generate autocorrelated time series
-  res <- stats::filter(rnorm(1000), filter = rep(1,5), circular = TRUE)
-  res[100:120] <- NA
-  (effAcf <- computeEffectiveAutoCorr(res))
-}
 
 
-#' @export
-varEffective <- function(
-  ### Estimate the variance of a correlated time series
-  res  ##<< numeric of autocorrelated numbers, usually observation - model residuals
-  , nEff = computeEffectiveNumObs(res, na.rm = na.rm) ##<< effective 
-  ## number of observations
-  , na.rm = FALSE  ##<< set to TRUE to remove NA cases before computation
-  , ...  ##<< further arguments to \code{\link{var}}
-) {
-  ##details<< The BLUE is not anymore the usual variance, but a modified
-  ## variance as given in \code{Zieba 2011}
-  a <- 1
-  var(res, na.rm = na.rm, ...) * nEff/(nEff - 1)
-  ### The estimated variance of the sample
+#' remove NA values at the start and end
+#'
+#' @param x numeric vector
+#'
+#' @return subset of x with leading and trailing NAs removed
+#' @keywords internal
+.trimNA <- function(x){
+  nisna <- complete.cases(x)
+  idx <- cumsum(nisna > 0) & rev(cumsum(rev(nisna))) > 0
+  x[idx]
 }
-attr(varEffective,"ex") <- function(){
-  # generate autocorrelated time series
-  res <- stats::filter(rnorm(1000), filter = rep(1,5), circular = TRUE)
-  res[100:120] <- NA
-  # if correlations are neglected, the estimate of the variance is biased low
-  (varNeglectCorr <- var(res, na.rm = TRUE))
-  (varCorr <- varEffective(res, na.rm = TRUE))
-}
-
 
 .tmp.f <- function(){
   dss <- dsfP %>% mutate(
@@ -147,13 +276,14 @@ attr(varEffective,"ex") <- function(){
   sd(dss$resp, na.rm = TRUE)/sqrt(nEff)
 }
 
+#' Construct the full correlation matrix from autocorrelation components.
+#'
+#' @param nRow number of rows in correlation matrix
+#' @param effAcf numeric vector of effective autocorrelation components
+#' . The first entry, which is defined as 1, is not used.
+#'
 #' @export
-getCorrMatFromAcf <- function(
-  ### Construct the full correlation matrix from autocorrelation components.
-  nRow      ##<< number of rows in correlation matrix
-  , effAcf  ##<< numeric vector of effective autocorrelation components
-  ##. The first entry, which is defined as 1, is not used.
-){
+getCorrMatFromAcf <- function(nRow, effAcf){
   nDiag <- length(effAcf) - 1
   if (nDiag < 1) return(Diagonal(nRow))
   bandSparse(
@@ -164,18 +294,20 @@ getCorrMatFromAcf <- function(
   )
 }
 
+#' set off-diagonal values of a matrix
+#'
+#' @param x numeric square matrix
+#' @param diag integer vector specifying the diagonals
+#'  0 is the center +1 the first row to upper and -2 the second row to lower
+#' @param value numeric vector of values to fill in
+#' @param isSymmetric  set to TRUE to to only
+#'  specify the upper diagonal element but also
+#'  set the lower in the mirrored diagonal
+#'
 #' @export
+#' @return matrix with modified diagonal elements
 setMatrixOffDiagonals <- function(
-  ### set off-diagonal values of the matrix
-  x        ##<< numeric square matrix
-  , diag = 1:length(value)  ##<< integer vector specifying the diagonals
-  ## 0 is the center +1 the first
-  ## row to upper and -2 the second row to lower
-  , value ##<< numeric vector of values to fill in
-  , isSymmetric = FALSE ##<< set to TRUE to to only
-  ## specify the upper diagonal element but also
-  ## set the lower in the mirrored diagonal
-){
+  x, diag = 1:length(value), value, isSymmetric = FALSE){
   if (!is.matrix(x) || !is.numeric(x) ) stop(
     "x must be a numeric matrix" )
   dimX <- dim(x)
@@ -195,64 +327,6 @@ setMatrixOffDiagonals <- function(
   }
   # dimensions may have gone lost
   dim(x) <- dimX
-  ##value<< matrix with modified diagonal elements
   x
-}
-
-seCorSqrtN <- function(
-  ### compute the standard error accounting for empirical correlations
-  x  ##<< numeric vector
-  , ...   ##<< further arguments to \code{\link{acf}}
-  , lag.max = round(sqrt(sum(is.finite(x))))  ##<< integer scalar:
-  ## maxium range of correlation
-){
-  # deprecated, superseeded by seCor
-  ##details<< computation according to 
-  ##  https://stats.stackexchange.com/questions/274635/calculating-error-of-mean-of-time-series
-  n <- length(x)  
-  kmax <- min(lag.max, n - 1)
-  varx <- var(x, na.rm = TRUE)
-  g1 <- varx * acf(x, lag.max = kmax, ..., plot = FALSE, na.action = na.pass)
-  g0 <- g1[1]
-  g <- g1[-1]
-  k <- 1:kmax
-  varCor <- 1/n*(g0 + 2*sum( (n - k)/n * g))
-  ##value<< numeric scalar of standard error of the mean of x
-  sqrt(varCor)
-}
-
-#' @export
-seCor <- function(
-  ### Compute the standard error accounting for empirical autocorrelations
-  x  ##<< numeric vector
-  , na.rm = FALSE ##<< logical. Should missing values be removed?
-  , effCov = ##<< numeric vector of effective covariance components
-    ## first entry is the variance. See \code{\link{computeEffectiveAutoCorr}}
-    computeEffectiveAutoCorr(x, type = "covariance")
-){
-  ##details<< Computation follows 
-  ## https://stats.stackexchange.com/questions/274635/calculating-error-of-mean-of-time-series.
-  ## 
-  ##details<< The default uses empirical autocorrelation
-  ## estimates from the supplied data up to first negative component.
-  ## For short series of \code{x} it is strongly recommended to to
-  ## provide \code{effCov} that was estimated on a longer time series.
-  n <- if (na.rm) length(na.omit(x)) else length(x)
-  #effAcf <- computeEffectiveAutoCorr(x, na.action = {if(na.rm) na.omit else na.pass})
-  # do not remove NAs for autocorrelation computation to preserve distances
-  if (n == 0) return(NA_real_)
-  g1 <- effCov[1:min(length(x),length(effCov))]
-  kmax <- length(g1) - 1
-  # if there is no empirical autocorrelation
-  if (kmax == 0) {
-    varx <- var(x, na.rm = na.rm)
-    return(sqrt(varx/n))
-  } 
-  g0 <- g1[1]
-  g <- g1[-1]
-  k <- 1:kmax
-  varCor <- 1/n*(g0 + 2*sum( (n - k)/n * g))
-  ##value<< numeric scalar of standard error of the mean of x
-  sqrt(varCor)
 }