+++ inst/doc/lognormalSum.R
@@ -1,8 +1,8 @@
-## ----eval=FALSE, include=FALSE-------------------------------------------
+## ----eval=FALSE, include=FALSE------------------------------------------------
 #  # twDev::genVigs()
-#  rmarkdown::render("lognormalSum.Rmd","md_document")
+#  #rmarkdown::render("lognormalSum.Rmd","md_document")
 
-## ----setup, include=FALSE------------------------------------------------
+## ----setup, include=FALSE-----------------------------------------------------
 library(knitr)
 opts_chunk$set(out.extra = 'style="display:block; margin: auto"'
     #, fig.align = "center"
@@ -21,13 +21,13 @@ knit_hooks$set(spar = function(before, options, envir) {
 })
 library(lognorm) 
 if (!require(ggplot2) || !require(dplyr) || !require(tidyr) || !require(purrr)) {
-	print("To generate this vignette, ggplo2, dplyr, tidyr, and purrr are required.")
-	exit(0)
+	print("To generate this vignette, ggplot2, dplyr, tidyr, and purrr are required.")
+	knit_exit()
 }
-themeTw <- theme_bw(base_size = 10) + 
+themeTw <- ggplot2::theme_bw(base_size = 10) + 
   theme(axis.title = element_text(size = 9))
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 # generate nSample values of two lognormal random variables
 mu1 = log(110)
 mu2 = log(100)
@@ -35,7 +35,7 @@ sigma1 = 0.25
 sigma2 = 0.15
 (coefSum <- estimateSumLognormal( c(mu1,mu2), c(sigma1,sigma2) ))
 
-## ---- echo=FALSE, fig.height=2.04, fig.width=3.27------------------------
+## ----densitySumTwo, echo=FALSE, fig.height=2.04, fig.width=3.27---------------
 nSample = 2000
 ds <- data.frame(
   x1 = rlnorm(nSample, mu1, sigma1)
@@ -56,6 +56,82 @@ themeTw +
 theme(legend.position = c(0.98,0.98), legend.justification = c(1,1)) +
 theme(axis.title.x = element_blank())
 
-## ----eval=FALSE, include=FALSE-------------------------------------------
-#  # complaint by CRAN: [md](aggregateCorrelated.md))
+## -----------------------------------------------------------------------------
+if (!requireNamespace("mvtnorm")) {
+  warning("Remainder of the vignette required mvtnorm installed.")
+  knitr::opts_chunk$set(error = TRUE) 
+}
+nObs <- 100; nRep <- 10000
+#nObs <- 1000; nRep <- 100
+xTrue <- rep(10, nObs)
+sigmaStar <- rep(1.7, nObs) # multiplicative stddev 
+theta <- getParmsLognormForExpval(xTrue, sigmaStar)
+# generate observations with correlated errors
+acf1 <- c(0.4,0.1)
+corrM <- setMatrixOffDiagonals(
+  diag(nrow = nObs), value = acf1, isSymmetric = TRUE)
+xObsN <- exp(mvtnorm::rmvnorm(
+  nRep, mean = theta[,1]
+  , sigma = diag(theta[,2]) %*% corrM %*% diag(theta[,2])))
+
+## ----draw100, echo=FALSE, fig.height=2.04, fig.width=3.27---------------------
+ds <- tibble(i = 1:nObs, xTrue, xObs = xObsN[1,], xErr =  xObs - xTrue)
+ggplot( ds, aes(i, xObs)) +
+  geom_line() +
+  geom_hline(yintercept = xTrue[1]) +
+  themeTw +   
+  theme(axis.title.x = element_blank())
+
+## ----echo=FALSE---------------------------------------------------------------
+c(1, acf1)
+
+## -----------------------------------------------------------------------------
+(effAcf <- computeEffectiveAutoCorr(ds$xErr))
+(nEff <- computeEffectiveNumObs(ds$xErr))
+
+## -----------------------------------------------------------------------------
+#coefSum <- estimateSumLognormal( theta[,1], theta[,2], effAcf = effAcf )
+coefSum <- estimateSumLognormal( theta[,1], theta[,2], effAcf = c(1,acf1) )
+setNames(exp(coefSum["sigma"]), "sigmaStar")
+
+## -----------------------------------------------------------------------------
+(sumExp <- getLognormMoments( coefSum[1], coefSum[2])[1,"mean"])
+
+## ----pdfSum100, echo=FALSE, fig.height=2.04, fig.width=3.27-------------------
+dsPredSum <- data.frame(
+  p = seq(0, 1, length.out = 100)[-c(1,100)] # percentiles
+) %>%
+  mutate( 
+    q = qlnorm(p, coefSum["mu"], coefSum["sigma"] )  # quantiles
+    ,density = dlnorm(q, coefSum["mu"], coefSum["sigma"])) # approximated density
+# density plot of the random draws
+ggplot(data.frame(y = rowSums(xObsN)), aes(y, color = "random draws")) + 
+  geom_density() +
+  # line plot of the lognorm density approximation
+  geom_line(data = dsPredSum, aes(q, density, color = "computed sum")) +
+  # expected value
+  geom_vline(xintercept = sumExp) +
+themeTw +
+theme(legend.position = c(0.98,0.98), legend.justification = c(1,1)) +
+theme(axis.title.x = element_blank()) +
+theme(legend.title = element_blank())
+
+## -----------------------------------------------------------------------------
+(coefMean <- setNames(c(coefSum["mu"] - log(nObs), coefSum["sigma"]), c("mu","sigma")))
+
+## ----pdfMean, echo=FALSE, fig.height=2.04, fig.width=3.27---------------------
+dsPredMean <- data.frame(
+  p = seq(0, 1, length.out = 100)[-c(1,100)] # percentiles
+) %>%
+  mutate( 
+    q = qlnorm(p, coefMean["mu"], coefMean["sigma"] )  # quantiles
+    ,density = dlnorm(q, coefMean["mu"], coefMean["sigma"])) # approximated density
+ggplot(data = dsPredMean, aes(q, density, color = "mean")) +
+  geom_line() +
+  geom_vline(xintercept = getLognormMoments(
+    coefMean["mu"],coefMean["sigma"])[1,"mean"]) +
+themeTw +
+theme(legend.position = c(0.98,0.98), legend.justification = c(1,1)) +
+theme(axis.title.x = element_blank()) +
+theme(legend.title = element_blank())